refactor: feed detail
