+++ src/app/feed/FeedItem.tsx
@@ -1,11 +1,11 @@
 import Link from 'next/link';
 import { useGetMyId } from '@/apis/member';
 import { type FeedItemType } from '@/apis/schema/feed';
-import HistoryThumbnail from '@/app/record/[id]/detail/HistoryThumbnail';
 import ReactionBar from '@/components/ReactionBar/ReactionBar';
 import Thumbnail from '@/components/Thumbnail/Thumbnail';
 import { EVENT_LOG_CATEGORY, EVENT_LOG_NAME } from '@/constants/eventLog';
 import { ROUTER } from '@/constants/router';
+import HistoryThumbnail from '@/pages/feed/HistoryThumbnail';
 import { eventLogger } from '@/utils';
 import { css } from '@styled-system/css';
 import dayjs from 'dayjs';
