+++ src/pages/record/[id]/follow.page.tsx
@@ -1,15 +1,23 @@
 import { Suspense } from 'react';
+import { type GetServerSidePropsContext } from 'next';
 import Header from '@/components/Header/Header';
 import MissionRecordDetail from '@/pages/record/[id]/detail/MissionRecordDetail';
 import { css } from '@styled-system/css';
 
-function FollowMissionRecordDetailPage() {
+export async function getServerSideProps(context: GetServerSidePropsContext) {
+  return {
+    props: {
+      id: context.query.id,
+    },
+  };
+}
+
+function FollowMissionRecordDetailPage({ id }: { id: string }) {
   return (
     <main className={mainWrapperCss}>
       <Header rightAction={'none'} title={'미션 내역'} />
       <Suspense>
-        {/* TODO : record id 잘 넣기 */}
-        <MissionRecordDetail isFollow recordId="1" />
+        <MissionRecordDetail isFollow recordId={id} />
       </Suspense>
     </main>
   );
